test(compiler): shard the pinned-corpus sweeps into parallel files
The evaluator-outcome sweep and the opt -passes=verify cross-check each
ran as one test walking the whole pinned corpus, which serialized
minutes of work inside a single vitest worker while the rest of the
machine sat idle — they were the two slowest tests under the parallel
gate (4m38 and 5m30). Each sweep now registers from four thin shard
files through one shared runner, with one it per program so a rank
report names the exact corpus entry that got slow.

The verifier sweep keeps its emitted-at-least-one guard per shard (a
shard that verified nothing is not a green shard), and per-program
ceilings are contention headroom, not performance assertions.

281 tests pass across the nine files; local wall time for the pair
drops from two serial multi-minute tests to shards the worker pool
packs freely.
