+++ packages/compiler/test/BootstrapEvaluationCorpus1.test.ts
@@ -1,0 +1,3 @@
+import { corpusOutcomeShard } from './support/corpusShards.js'
+
+corpusOutcomeShard(1)
