+++ packages/compiler/test/BootstrapEvaluationCorpus2.test.ts
@@ -1,0 +1,3 @@
+import { corpusOutcomeShard } from './support/corpusShards.js'
+
+corpusOutcomeShard(2)
