+++ packages/compiler/test/BootstrapEvaluationCorpus3.test.ts
@@ -1,0 +1,3 @@
+import { corpusOutcomeShard } from './support/corpusShards.js'
+
+corpusOutcomeShard(3)
