+++ packages/compiler/test/BootstrapEvaluationCorpus4.test.ts
@@ -1,0 +1,3 @@
+import { corpusOutcomeShard } from './support/corpusShards.js'
+
+corpusOutcomeShard(4)
