+++ packages/compiler/test/ModuleVerification1.test.ts
@@ -1,0 +1,3 @@
+import { moduleVerificationShard } from './support/corpusShards.js'
+
+moduleVerificationShard(1)
