+++ packages/compiler/test/ModuleVerification3.test.ts
@@ -1,0 +1,3 @@
+import { moduleVerificationShard } from './support/corpusShards.js'
+
+moduleVerificationShard(3)
